Keyboard shortcuts for changing volume and pan are now also supported for midi tracks in arranger. Accelerator key (shift) isn't supported right now since it was simply too hard to find sensible shortcuts that make sense, without using shift
Basically the following shortcuts can be used:
Ctrl+Shift+Arrow keys adjust volume (up/down) pan (left/right) in both
arranger and mixer.
Switching between tracks is done with
Ctrl+Up/Down in tracklist
and
Ctrl+Alt+Left/Right in mixer

The shortcuts can be changed in the shortcut dialog.

Reminder for an old but useful function: In the mixer it can happen
that focus is stuck on a child widget, pressing ESC will move focus
to the mixer as a whole.